CMDB_SYSTEM/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.5
djangorestframework==3.7.7
idna==2.6
Naked==0.1.31
pycrypto==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
urllib3==1.22
